refactor: restructure fakehuman directory
