+++ src/core/execution/ExecutionManager.ts
@@ -15,7 +15,7 @@ import { DonateGoldExecution } from "./DonateGoldExecution";
 import { DonateTroopsExecution } from "./DonateTroopExecution";
 import { EmbargoExecution } from "./EmbargoExecution";
 import { EmojiExecution } from "./EmojiExecution";
-import { FakeHumanCoordinator } from "./fakehuman/FakeHumanCoordinator";
+import { FakeHumanCoordinator } from "./fakehuman/core/FakeHumanCoordinator";
 import { MarkDisconnectedExecution } from "./MarkDisconnectedExecution";
 import { MoveWarshipExecution } from "./MoveWarshipExecution";
 import { NoOpExecution } from "./NoOpExecution";
